backend/requirements.txt
aioredis==1.3.0
amqp==2.5.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
billiard==3.6.1.0
cachetools==3.1.1
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2.6
django-appconf==1.0.3
django-cleanup==4.0.0
django-cors-headers==3.1.1
django-extra-fields==1.2.4
django-googledrive-storage==1.3.5
django-templated-mail==1.1.1
djangorestframework==3.10.3
djoser==2.0.3
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
gunicorn==20.0.0
hiredis==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
kombu==4.6.5
more-itertools==7.2.0
msgpack==0.6.2
oauth2client==4.1.3
Pillow==6.2.0
pip-autoremove==0.9.1
psycopg2==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
redis==2.10.6
requests==2.22.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
whitenoise==4.1.4
zipp==0.6.0
zope.interface==4.7.1
